requirements.txt
fastapi
uvicorn
python-dotenv
supabase
python-jose
pydantic
langchain
langchain-community
langchain-groq
chromadb
sentence-transformers
groq
